// interfaces/ITable.ts
export interface ITable {
    id?: number;
    card1: string;
    card2: string;
    card3: string;
    card4: string;
    card5: string;
    card6: string;
    card7: string;
    card8: string;
    card9: string;
    card10: string;
    card11: string;
    card12: string;
    card13: string;
    card14: string;
    card15: string;
    card16: string;
}